Simplify metadata-logging interface.
