+++ app/gui/src/model/project/synchronized.rs
@@ -32,15 +32,7 @@ use parser::Parser;
 // === Profiling ===
 // =================
 
-thread_local! {
-    /// Profiling metadata logger for RPC event names..
-    static RPC_EVENT_LOGGER: enso_profiler::MetadataLogger<& 'static str> = enso_profiler::MetadataLogger::new("RpcEvent");
-}
-
-/// Log an RPC Event to the profiling framework.
-pub fn log_rpc_event(event_name: &'static str) {
-    RPC_EVENT_LOGGER.with(|logger| logger.log(event_name));
-}
+profiler::metadata_logger!("RpcEvent", log_rpc_event(&'static str));
